requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.11.0
#appscript==1.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
billiard==4.2.2
Brotli==1.1.0
cachetools==6.2.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cobble==0.1.4
constantly==23.10.4
cryptography==45.0.4
cssselect2==0.8.0
daphne==4.2.0
Django==5.2.3
django-cors-headers==4.7.0
django-crontab==0.7.1
django-crum==0.7.9
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
fonttools==4.59.2
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipinfo==5.2.1
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.1
mammoth==1.10.0
Markdown==3.9
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
propcache==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
reportlab==4.4.3
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.37.0
#uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
zopfli==0.2.3.post1
